Move data-related classes into their own package
