+++ src/main/java/com/recurse/portfolio/data/RecurseProfile.java
@@ -1,4 +1,4 @@
-package com.recurse.portfolio;
+package com.recurse.portfolio.data;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
