+++ src/main/java/com/recurse/portfolio/data/UserRepository.java
@@ -1,11 +1,11 @@
-package com.recurse.portfolio;
+package com.recurse.portfolio.data;
 
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-interface UserRepository
+public interface UserRepository
     extends CrudRepository<User, Integer>
 {
     @Query("select * from users where recurse_profile_id = :recurseProfileId")
